add : 타이머 로직 수정 및 알림 기능 추가 (레시피 페이지에서만 보이게 수정)
